--- docker-testkit/src/main/scala/stormlantern/dockertestkit/orchestration/Orchestration.scala
@@ -1,5 +1,0 @@
-package stormlantern.dockertestkit.orchestration
-
-class Orchestration {
-
-}
